solana-ctf: Create per event USDC ata
With this change we are introducing a major security update for
our solana-ctf program. All the escrow account which is going to
hold user fund are now per event based. All these escrow accounts
are now PDA derived instead of ARKA owned. This increases the
overall security of the protocol because ARKA does not own these
escrow accounts anymore and they are all controlled by the solana
-ctf program instead.

Signed-off-by: Jinank Jain <jinank@futurestech.xyz>